refactor: extract command handlers and message routing out of bot.py
Move topic_closed_handler and topic_edited_handler to topic_lifecycle.py,
unbind_command to cleanup.py, recall_command to command_history.py,
screenshot_command and panes_command to screenshot_callbacks.py, and
handle_new_message (with _MIN_THINKING_LENGTH) to a new message_routing.py
module. Update all test import paths accordingly.
