+++ src/ccgram/bot.py
@@ -21,14 +21,9 @@ import asyncio
 import contextlib
 import structlog
 import os
-import re
 import signal
-from pathlib import Path
 
 from telegram import (
-    Bot,
-    InlineKeyboardButton,
-    InlineKeyboardMarkup,
     InlineQueryResultArticle,
     InputTextMessageContent,
     Update,
@@ -64,9 +59,6 @@ from .handlers.command_orchestration import (
     sync_scoped_provider_menu as _sync_scoped_provider_menu,
     setup_menu_refresh_job,
 )
-from .handlers.callback_data import (
-    CB_PANE_SCREENSHOT,
-)
 from .handlers.callback_helpers import get_thread_id as _get_thread_id
 from .handlers.callback_registry import dispatch as _dispatch_callback
 from .handlers.callback_registry import load_handlers as _load_callback_handlers
@@ -74,52 +66,33 @@ from .handlers.restore_command import restore_command
 from .handlers.resume_command import resume_command
 from .handlers.send_command import send_command
 from .handlers.directory_browser import clear_browse_state
-from .handlers.cleanup import clear_topic_state
-from .handlers.topic_emoji import strip_emoji_prefix, update_stored_topic_name
+from .handlers.cleanup import unbind_command
+from .handlers.command_history import recall_command
+from .handlers.message_routing import handle_new_message
+from .handlers.screenshot_callbacks import panes_command, screenshot_command
+from .handlers.topic_lifecycle import topic_closed_handler, topic_edited_handler
 from .handlers.history import send_history
 from .handlers.sessions_dashboard import sessions_command
 from .handlers.sync_command import sync_command
 from .handlers.upgrade import upgrade_command
-from .handlers.interactive_ui import (
-    INTERACTIVE_TOOL_NAMES,
-    clear_interactive_mode,
-    clear_interactive_msg,
-    get_interactive_msg_id,
-    handle_interactive_ui,
-    set_interactive_mode,
-)
 from .handlers.message_queue import (
-    enqueue_content_message,
-    enqueue_status_update,
-    get_message_queue,
     shutdown_workers,
 )
 from .handlers.message_sender import safe_reply
-from .handlers.response_builder import build_response_parts
 from .handlers.polling_coordinator import status_poll_loop
 from .handlers.file_handler import handle_document_message, handle_photo_message
 from .handlers.voice_handler import handle_voice_message
 from .handlers.text_handler import handle_text_message
 from .session import session_manager
-from .user_preferences import user_preferences
 from .session_monitor import NewMessage, NewWindowEvent, SessionMonitor
 from .thread_router import thread_router
 from .telegram_request import ResilientPollingHTTPXRequest
-from .tmux_manager import tmux_manager
 from .utils import handle_general_topic_message, is_general_topic, task_done_callback
 
 logger = structlog.get_logger()
 
 # Error keyword pattern for errors_only notification mode (word boundaries)
-_ERROR_KEYWORDS_RE = re.compile(
-    r"\b(?:error|exception|failed|traceback|stderr|assertion)\b", re.IGNORECASE
-)
 
-# Max label length for /recall command buttons (wider than status bar buttons)
-_RECALL_LABEL_MAX = 40
-# Minimum length for thinking content to be forwarded to Telegram.
-# Shorter thinking blocks (e.g. bare "(thinking)") are noise and get skipped.
-_MIN_THINKING_LENGTH = 20
 # Session monitor instance
 session_monitor: SessionMonitor | None = None
 
@@ -211,326 +184,6 @@ async def commands_command(update: Update, _context: ContextTypes.DEFAULT_TYPE) 
         original = cmd.name if cmd.name.startswith("/") else f"/{cmd.name}"
         lines.append(f"- `/{cmd.telegram_name}` \u2192 `{original}`")
     await safe_reply(update.message, "\n".join(lines))
-
-
-async def topic_closed_handler(
-    update: Update, context: ContextTypes.DEFAULT_TYPE
-) -> None:
-    """Handle topic closure — unbind thread but keep the tmux window alive.
-
-    The window becomes "unbound" and is available for rebinding via the window
-    picker when a new topic is created. Unbound windows are auto-killed after
-    the configured TTL (autoclose_done_minutes) by the status polling loop.
-    """
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        return
-
-    window_id = thread_router.get_window_for_thread(user.id, thread_id)
-    if window_id:
-        display = thread_router.get_display_name(window_id)
-        # Clean up BEFORE unbind — resolve_chat_id needs group_chat_ids
-        # which unbind_thread deletes.  window_dead=False because the
-        # window stays alive for rebinding.
-        await clear_topic_state(
-            user.id,
-            thread_id,
-            context.bot,
-            context.user_data,
-            window_id=window_id,
-            window_dead=False,
-        )
-        thread_router.unbind_thread(user.id, thread_id)
-        logger.info(
-            "Topic closed: window %s unbound (kept alive for rebinding, user=%d, thread=%d)",
-            display,
-            user.id,
-            thread_id,
-        )
-    else:
-        logger.debug(
-            "Topic closed: no binding (user=%d, thread=%d)", user.id, thread_id
-        )
-
-
-async def topic_edited_handler(
-    update: Update, _context: ContextTypes.DEFAULT_TYPE
-) -> None:
-    """Handle topic rename — sync new name to tmux window and emoji cache.
-
-    Ignores icon-only edits (name is None) and emoji-only changes from the bot
-    itself (clean name unchanged after stripping prefixes).
-    """
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-    if not update.message or not update.message.forum_topic_edited:
-        return
-
-    new_name = update.message.forum_topic_edited.name
-    if not new_name:
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        return
-
-    chat_id = update.effective_chat.id if update.effective_chat else None
-    if chat_id is None:
-        return
-
-    window_id = thread_router.get_window_for_chat_thread(chat_id, thread_id)
-    if not window_id:
-        logger.debug("Topic edited: no binding (thread=%d)", thread_id)
-        return
-
-    clean_name = strip_emoji_prefix(new_name)
-
-    # Loop guard: if clean name matches current display name, this was a
-    # bot-originated emoji/mode change — skip to prevent rename loops.
-    current_display = thread_router.get_display_name(window_id)
-    if current_display and strip_emoji_prefix(current_display) == clean_name:
-        logger.debug(
-            "Topic edited: name unchanged after strip, skipping (thread=%d)", thread_id
-        )
-        return
-
-    renamed = await tmux_manager.rename_window(window_id, clean_name)
-    if renamed:
-        session_manager.set_display_name(window_id, clean_name)
-        update_stored_topic_name(chat_id, thread_id, clean_name)
-        logger.info(
-            "Topic renamed: window %s → %r (thread=%d)",
-            window_id,
-            clean_name,
-            thread_id,
-        )
-
-
-async def unbind_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Disconnect a topic from its tmux window without killing the session."""
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-    if not update.message:
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        if (
-            update.message
-            and update.effective_chat
-            and is_general_topic(update.message)
-        ):
-            await handle_general_topic_message(
-                update.get_bot(), update.message, update.effective_chat.id
-            )
-        else:
-            await safe_reply(update.message, "\u274c Use this command inside a topic.")
-        return
-
-    window_id = thread_router.get_window_for_thread(user.id, thread_id)
-    if not window_id:
-        await safe_reply(
-            update.message, "\u274c This topic is not bound to any session."
-        )
-        return
-
-    display = thread_router.get_display_name(window_id)
-    # Enqueue a status clear to actually delete the Telegram status message
-    # (clear_topic_state only clears the tracking dict, leaving a ghost)
-    await enqueue_status_update(context.bot, user.id, window_id, None, thread_id)
-    await clear_topic_state(
-        user.id,
-        thread_id,
-        context.bot,
-        context.user_data,
-        window_id=window_id,
-        window_dead=False,
-    )
-    thread_router.unbind_thread(user.id, thread_id)
-    await safe_reply(
-        update.message,
-        f"\u2702 Unbound from window `{display}`. The session is still running.\n"
-        "Send a message in this topic to rebind or create a new session.",
-    )
-
-
-async def screenshot_command(
-    update: Update, _context: ContextTypes.DEFAULT_TYPE
-) -> None:
-    """Capture and send a terminal screenshot for the current topic."""
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-    if not update.message:
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        if (
-            update.message
-            and update.effective_chat
-            and is_general_topic(update.message)
-        ):
-            await handle_general_topic_message(
-                update.get_bot(), update.message, update.effective_chat.id
-            )
-        else:
-            await safe_reply(update.message, "\u274c Use this command inside a topic.")
-        return
-
-    window_id = thread_router.get_window_for_thread(user.id, thread_id)
-    if not window_id:
-        await safe_reply(
-            update.message, "\u274c This topic is not bound to any session."
-        )
-        return
-
-    w = await tmux_manager.find_window_by_id(window_id)
-    if not w:
-        await safe_reply(update.message, "\u274c Window no longer exists.")
-        return
-
-    pane_text = await tmux_manager.capture_pane(w.window_id, with_ansi=True)
-    if not pane_text:
-        await safe_reply(update.message, "\u274c Failed to capture terminal.")
-        return
-
-    import io
-
-    from .handlers.screenshot_callbacks import build_screenshot_keyboard
-    from .screenshot import text_to_image
-
-    png_bytes = await text_to_image(pane_text, with_ansi=True)
-    keyboard = build_screenshot_keyboard(window_id)
-    chat_id = thread_router.resolve_chat_id(user.id, thread_id)
-    try:
-        await update.message.get_bot().send_document(
-            chat_id=chat_id,
-            document=io.BytesIO(png_bytes),
-            filename="screenshot.png",
-            reply_markup=keyboard,
-            message_thread_id=thread_id,
-        )
-    except TelegramError as e:
-        logger.error("Failed to send screenshot: %s", e)
-        await safe_reply(update.message, "\u274c Failed to send screenshot.")
-
-
-async def panes_command(update: Update, _context: ContextTypes.DEFAULT_TYPE) -> None:
-    """List all panes in the current topic's window."""
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-    if not update.message:
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        if (
-            update.message
-            and update.effective_chat
-            and is_general_topic(update.message)
-        ):
-            await handle_general_topic_message(
-                update.get_bot(), update.message, update.effective_chat.id
-            )
-        else:
-            await safe_reply(update.message, "\u274c Use this command inside a topic.")
-        return
-
-    window_id = thread_router.get_window_for_thread(user.id, thread_id)
-    if not window_id:
-        await safe_reply(
-            update.message, "\u274c This topic is not bound to any session."
-        )
-        return
-
-    panes = await tmux_manager.list_panes(window_id)
-    if len(panes) <= 1:
-        await safe_reply(
-            update.message,
-            "\U0001f4d0 Single pane \u2014 no multi-pane layout detected.",
-        )
-        return
-
-    from .handlers.polling_strategies import has_pane_alert
-
-    lines = [f"\U0001f4d0 {len(panes)} panes in window\n"]
-    buttons: list[InlineKeyboardButton] = []
-    for pane in panes:
-        prefix = "\U0001f4cd" if pane.active else "  "
-        label = f"Pane {pane.index} ({pane.command})"
-        suffix_parts: list[str] = []
-        if pane.active:
-            suffix_parts.append("active")
-        if has_pane_alert(pane.pane_id):
-            prefix = "\u26a0\ufe0f"
-            suffix_parts.append("blocked")
-        elif not pane.active:
-            suffix_parts.append("running")
-        suffix = f" \u2014 {', '.join(suffix_parts)}" if suffix_parts else ""
-        lines.append(f"{prefix} {label}{suffix}")
-        buttons.append(
-            InlineKeyboardButton(
-                f"\U0001f4f7 {pane.index}",
-                callback_data=f"{CB_PANE_SCREENSHOT}{window_id}:{pane.pane_id}"[:64],
-            )
-        )
-
-    keyboard = InlineKeyboardMarkup([buttons]) if buttons else None
-    await safe_reply(update.message, "\n".join(lines), reply_markup=keyboard)
-
-
-async def recall_command(update: Update, _context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Show recent command history for the current topic."""
-    user = update.effective_user
-    if not user or not is_user_allowed(user.id):
-        return
-    if not update.message:
-        return
-
-    thread_id = _get_thread_id(update)
-    if thread_id is None:
-        if (
-            update.message
-            and update.effective_chat
-            and is_general_topic(update.message)
-        ):
-            await handle_general_topic_message(
-                update.get_bot(), update.message, update.effective_chat.id
-            )
-        else:
-            await safe_reply(update.message, "\u274c Use this command inside a topic.")
-        return
-
-    from .handlers.command_history import (
-        INLINE_QUERY_MAX,
-        get_history,
-        truncate_for_display,
-    )
-
-    history = get_history(user.id, thread_id, limit=10)
-    if not history:
-        await safe_reply(update.message, "\U0001f4cb No command history yet.")
-        return
-
-    rows = []
-    for cmd in history:
-        label = truncate_for_display(cmd, _RECALL_LABEL_MAX)
-        query = cmd[:INLINE_QUERY_MAX]
-        rows.append(
-            [InlineKeyboardButton(label, switch_inline_query_current_chat=query)]
-        )
-    keyboard = InlineKeyboardMarkup(rows)
-    await safe_reply(
-        update.message, "\U0001f4cb Recent commands:", reply_markup=keyboard
-    )
 
 
 async def toolbar_command(update: Update, _context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -673,128 +326,6 @@ async def text_handler(update: Update, context: ContextTypes.DEFAULT_TYPE) -> No
 
     await _sync_scoped_menu_for_text_context(update, user.id)
     await handle_text_message(update, context)
-
-
-# --- Streaming response / notifications ---
-
-
-async def handle_new_message(msg: NewMessage, bot: Bot) -> None:
-    """Handle a new assistant message — enqueue for sequential processing.
-
-    Messages are queued per-user to ensure status messages always appear last.
-    Routes via thread_bindings to deliver to the correct topic.
-    """
-    status = "complete" if msg.is_complete else "streaming"
-    logger.info(
-        "handle_new_message [%s]: session=%s, text_len=%d",
-        status,
-        msg.session_id,
-        len(msg.text),
-    )
-
-    # Find users whose thread-bound window matches this session
-    active_users = session_manager.find_users_for_session(msg.session_id)
-
-    if not active_users:
-        logger.info("No active users for session %s", msg.session_id)
-        return
-
-    for user_id, window_id, thread_id in active_users:
-        structlog.contextvars.clear_contextvars()
-        structlog.contextvars.bind_contextvars(
-            window_id=window_id, session_id=msg.session_id
-        )
-        # Check notification mode — skip suppressed messages.
-        # All tool_use/tool_result MUST pass through regardless of mode: the message
-        # queue edits tool_use messages in-place when tool_result arrives, so filtering
-        # one half would break pairing and leave orphaned messages. This means muted/
-        # errors_only sessions still deliver tool flow — an accepted trade-off.
-        notif_mode = session_manager.get_notification_mode(window_id)
-        is_tool_flow = msg.tool_name in INTERACTIVE_TOOL_NAMES or msg.content_type in (
-            "tool_use",
-            "tool_result",
-        )
-        if not is_tool_flow:
-            if notif_mode == "muted":
-                continue
-            if notif_mode == "errors_only" and not _ERROR_KEYWORDS_RE.search(
-                msg.text or ""
-            ):
-                continue
-
-        # Skip trivial thinking messages — "(thinking)" or very short thinking
-        # blocks carry no useful reasoning content. Only forward thinking with
-        # substantial text (>= 20 chars after stripping).
-        if msg.content_type == "thinking":
-            stripped = (msg.text or "").strip()
-            if len(stripped) < _MIN_THINKING_LENGTH:
-                continue
-
-        # Handle interactive tools specially - capture terminal and send UI
-        if msg.tool_name in INTERACTIVE_TOOL_NAMES and msg.content_type == "tool_use":
-            # Mark interactive mode BEFORE sleeping so polling skips this window
-            set_interactive_mode(user_id, window_id, thread_id)
-            # Flush pending messages (e.g. plan content) before sending interactive UI
-            queue = get_message_queue(user_id)
-            if queue:
-                await queue.join()
-            # Wait briefly for Claude Code to render the question UI
-            await asyncio.sleep(0.3)
-            handled = await handle_interactive_ui(bot, user_id, window_id, thread_id)
-            if handled:
-                # Update user's read offset
-                session = await session_manager.resolve_session_for_window(window_id)
-                if session and session.file_path:
-                    try:
-                        file_size = Path(session.file_path).stat().st_size
-                        user_preferences.update_user_window_offset(
-                            user_id, window_id, file_size
-                        )
-                    except OSError:
-                        pass
-                continue  # Don't send the normal tool_use message
-            else:
-                # UI not rendered — clear the early-set mode
-                clear_interactive_mode(user_id, thread_id)
-
-        # Any non-interactive message means the interaction is complete — delete the UI message
-        if get_interactive_msg_id(user_id, thread_id):
-            await clear_interactive_msg(user_id, bot, thread_id)
-
-        parts = build_response_parts(
-            msg.text,
-            msg.is_complete,
-            msg.content_type,
-            msg.role,
-        )
-
-        if msg.is_complete:
-            # Enqueue content message task
-            # Note: tool_result editing is handled inside _process_content_task
-            # to ensure sequential processing with tool_use message sending
-            await enqueue_content_message(
-                bot=bot,
-                user_id=user_id,
-                window_id=window_id,
-                parts=parts,
-                tool_use_id=msg.tool_use_id,
-                tool_name=msg.tool_name,
-                content_type=msg.content_type,
-                text=msg.text,
-                thread_id=thread_id,
-            )
-
-            # Update user's read offset to current file position
-            # This marks these messages as "read" for this user
-            session = await session_manager.resolve_session_for_window(window_id)
-            if session and session.file_path:
-                try:
-                    file_size = Path(session.file_path).stat().st_size
-                    user_preferences.update_user_window_offset(
-                        user_id, window_id, file_size
-                    )
-                except OSError:
-                    pass
 
 
 # --- App lifecycle ---
