+++ src/ccgram/handlers/cleanup.py
@@ -9,9 +9,15 @@ Functions:
   - clear_topic_state: Clean up all memory state for a specific topic
 """
 
-from typing import Any
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
 
 from telegram import Bot
+
+if TYPE_CHECKING:
+    from telegram import Update
+    from telegram.ext import ContextTypes
 
 from ..utils import log_throttle_reset
 from .interactive_ui import clear_interactive_msg
@@ -94,3 +100,57 @@ async def clear_topic_state(
         stale = [k for k in voice_store if k[0] == chat_id]
         for k in stale:
             voice_store.pop(k, None)
+
+
+async def unbind_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    """Disconnect a topic from its tmux window without killing the session."""
+    from ..config import config
+    from ..thread_router import thread_router
+    from ..utils import handle_general_topic_message, is_general_topic
+    from .callback_helpers import get_thread_id
+    from .message_queue import enqueue_status_update
+    from .message_sender import safe_reply
+
+    user = update.effective_user
+    if not user or not config.is_user_allowed(user.id):
+        return
+    if not update.message:
+        return
+
+    thread_id = get_thread_id(update)
+    if thread_id is None:
+        if (
+            update.message
+            and update.effective_chat
+            and is_general_topic(update.message)
+        ):
+            await handle_general_topic_message(
+                update.get_bot(), update.message, update.effective_chat.id
+            )
+        else:
+            await safe_reply(update.message, "\u274c Use this command inside a topic.")
+        return
+
+    window_id = thread_router.get_window_for_thread(user.id, thread_id)
+    if not window_id:
+        await safe_reply(
+            update.message, "\u274c This topic is not bound to any session."
+        )
+        return
+
+    display = thread_router.get_display_name(window_id)
+    await enqueue_status_update(context.bot, user.id, window_id, None, thread_id)
+    await clear_topic_state(
+        user.id,
+        thread_id,
+        context.bot,
+        context.user_data,
+        window_id=window_id,
+        window_dead=False,
+    )
+    thread_router.unbind_thread(user.id, thread_id)
+    await safe_reply(
+        update.message,
+        f"\u2702 Unbound from window `{display}`. The session is still running.\n"
+        "Send a message in this topic to rebind or create a new session.",
+    )
