+++ src/ccgram/handlers/command_history.py
@@ -10,9 +10,16 @@ Functions:
   - truncate_for_display: Truncate text for button labels
 """
 
+from __future__ import annotations
+
 from collections import deque
+from typing import TYPE_CHECKING
 
 from ..topic_state_registry import topic_state
+
+if TYPE_CHECKING:
+    from telegram import Update
+    from telegram.ext import ContextTypes
 
 HISTORY_MAX = 20
 
@@ -62,3 +69,56 @@ def truncate_for_display(text: str, max_len: int) -> str:
     if len(text) <= max_len:
         return text
     return text[: max_len - 1] + "\u2026"
+
+
+_RECALL_LABEL_MAX = 40
+
+
+async def recall_command(
+    update: Update,
+    _context: ContextTypes.DEFAULT_TYPE,
+) -> None:
+    """Show recent command history for the current topic."""
+    from telegram import InlineKeyboardButton, InlineKeyboardMarkup
+
+    from ..config import config
+    from ..utils import handle_general_topic_message, is_general_topic
+    from .callback_helpers import get_thread_id
+    from .message_sender import safe_reply
+
+    user = update.effective_user
+    if not user or not config.is_user_allowed(user.id):
+        return
+    if not update.message:
+        return
+
+    thread_id = get_thread_id(update)
+    if thread_id is None:
+        if (
+            update.message
+            and update.effective_chat
+            and is_general_topic(update.message)
+        ):
+            await handle_general_topic_message(
+                update.get_bot(), update.message, update.effective_chat.id
+            )
+        else:
+            await safe_reply(update.message, "\u274c Use this command inside a topic.")
+        return
+
+    history = get_history(user.id, thread_id, limit=10)
+    if not history:
+        await safe_reply(update.message, "\U0001f4cb No command history yet.")
+        return
+
+    rows = []
+    for cmd in history:
+        label = truncate_for_display(cmd, _RECALL_LABEL_MAX)
+        query = cmd[:INLINE_QUERY_MAX]
+        rows.append(
+            [InlineKeyboardButton(label, switch_inline_query_current_chat=query)]
+        )
+    keyboard = InlineKeyboardMarkup(rows)
+    await safe_reply(
+        update.message, "\U0001f4cb Recent commands:", reply_markup=keyboard
+    )
