+++ src/ccgram/handlers/screenshot_callbacks.py
@@ -585,6 +585,146 @@ def _schedule_key_refresh(
     _pending_key_refreshes[refresh_key] = asyncio.create_task(_do_refresh())
 
 
+# ------------------------------------------------------------------
+# Command handlers (moved from bot.py)
+# ------------------------------------------------------------------
+
+
+async def screenshot_command(
+    update: Update, _context: ContextTypes.DEFAULT_TYPE
+) -> None:
+    """Capture and send a terminal screenshot for the current topic."""
+    from ..config import config
+    from ..utils import handle_general_topic_message, is_general_topic
+    from .message_sender import safe_reply
+
+    user = update.effective_user
+    if not user or not config.is_user_allowed(user.id):
+        return
+    if not update.message:
+        return
+
+    thread_id = get_thread_id(update)
+    if thread_id is None:
+        if (
+            update.message
+            and update.effective_chat
+            and is_general_topic(update.message)
+        ):
+            await handle_general_topic_message(
+                update.get_bot(), update.message, update.effective_chat.id
+            )
+        else:
+            await safe_reply(update.message, "\u274c Use this command inside a topic.")
+        return
+
+    window_id = thread_router.get_window_for_thread(user.id, thread_id)
+    if not window_id:
+        await safe_reply(
+            update.message, "\u274c This topic is not bound to any session."
+        )
+        return
+
+    w = await tmux_manager.find_window_by_id(window_id)
+    if not w:
+        await safe_reply(update.message, "\u274c Window no longer exists.")
+        return
+
+    pane_text = await tmux_manager.capture_pane(w.window_id, with_ansi=True)
+    if not pane_text:
+        await safe_reply(update.message, "\u274c Failed to capture terminal.")
+        return
+
+    import io
+
+    png_bytes = await text_to_image(pane_text, with_ansi=True)
+    keyboard = build_screenshot_keyboard(window_id)
+    chat_id = thread_router.resolve_chat_id(user.id, thread_id)
+    try:
+        await update.message.get_bot().send_document(
+            chat_id=chat_id,
+            document=io.BytesIO(png_bytes),
+            filename="screenshot.png",
+            reply_markup=keyboard,
+            message_thread_id=thread_id,
+        )
+    except TelegramError as e:
+        logger.error("Failed to send screenshot: %s", e)
+        await safe_reply(update.message, "\u274c Failed to send screenshot.")
+
+
+async def panes_command(update: Update, _context: ContextTypes.DEFAULT_TYPE) -> None:  # noqa: C901
+    """List all panes in the current topic's window."""
+    from telegram import InlineKeyboardButton, InlineKeyboardMarkup
+
+    from ..config import config
+    from ..utils import handle_general_topic_message, is_general_topic
+    from .callback_data import CB_PANE_SCREENSHOT
+    from .message_sender import safe_reply
+
+    user = update.effective_user
+    if not user or not config.is_user_allowed(user.id):
+        return
+    if not update.message:
+        return
+
+    thread_id = get_thread_id(update)
+    if thread_id is None:
+        if (
+            update.message
+            and update.effective_chat
+            and is_general_topic(update.message)
+        ):
+            await handle_general_topic_message(
+                update.get_bot(), update.message, update.effective_chat.id
+            )
+        else:
+            await safe_reply(update.message, "\u274c Use this command inside a topic.")
+        return
+
+    window_id = thread_router.get_window_for_thread(user.id, thread_id)
+    if not window_id:
+        await safe_reply(
+            update.message, "\u274c This topic is not bound to any session."
+        )
+        return
+
+    panes = await tmux_manager.list_panes(window_id)
+    if len(panes) <= 1:
+        await safe_reply(
+            update.message,
+            "\U0001f4d0 Single pane \u2014 no multi-pane layout detected.",
+        )
+        return
+
+    from .polling_strategies import has_pane_alert
+
+    lines = [f"\U0001f4d0 {len(panes)} panes in window\n"]
+    buttons: list[InlineKeyboardButton] = []
+    for pane in panes:
+        prefix = "\U0001f4cd" if pane.active else "  "
+        label = f"Pane {pane.index} ({pane.command})"
+        suffix_parts: list[str] = []
+        if pane.active:
+            suffix_parts.append("active")
+        if has_pane_alert(pane.pane_id):
+            prefix = "\u26a0\ufe0f"
+            suffix_parts.append("blocked")
+        elif not pane.active:
+            suffix_parts.append("running")
+        suffix = f" \u2014 {', '.join(suffix_parts)}" if suffix_parts else ""
+        lines.append(f"{prefix} {label}{suffix}")
+        buttons.append(
+            InlineKeyboardButton(
+                f"\U0001f4f7 {pane.index}",
+                callback_data=f"{CB_PANE_SCREENSHOT}{window_id}:{pane.pane_id}"[:64],
+            )
+        )
+
+    keyboard = InlineKeyboardMarkup([buttons]) if buttons else None
+    await safe_reply(update.message, "\n".join(lines), reply_markup=keyboard)
+
+
 # --- Registry dispatch entry point ---
