+++ tests/ccgram/test_thinking_filter.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ccgram.bot import _MIN_THINKING_LENGTH
+from ccgram.handlers.message_routing import _MIN_THINKING_LENGTH
 
 
 def _should_skip_thinking(text: str | None) -> bool:
