+++ tests/ccgram/test_topic_close.py
@@ -11,17 +11,17 @@ def _make_update(thread_id: int = 42, user_id: int = 1) -> MagicMock:
     return update
 
 
-_PATCH_ALLOWED = patch("ccgram.bot.is_user_allowed", return_value=True)
+_PATCH_ALLOWED = patch("ccgram.config.Config.is_user_allowed", return_value=True)
 
 
 class TestTopicClosedHandler:
     @_PATCH_ALLOWED
-    @patch("ccgram.bot.clear_topic_state", new_callable=AsyncMock)
-    @patch("ccgram.bot.thread_router")
+    @patch("ccgram.handlers.topic_lifecycle.clear_topic_state", new_callable=AsyncMock)
+    @patch("ccgram.handlers.topic_lifecycle.thread_router")
     async def test_unbinds_bound_topic(
         self, mock_tr: MagicMock, mock_clear: AsyncMock, _allowed: MagicMock
     ) -> None:
-        from ccgram.bot import topic_closed_handler
+        from ccgram.handlers.topic_lifecycle import topic_closed_handler
 
         mock_tr.get_window_for_thread.return_value = "@0"
         mock_tr.get_display_name.return_value = "my-project"
@@ -42,12 +42,12 @@ class TestTopicClosedHandler:
         mock_tr.unbind_thread.assert_called_once_with(1, 42)
 
     @_PATCH_ALLOWED
-    @patch("ccgram.bot.clear_topic_state", new_callable=AsyncMock)
-    @patch("ccgram.bot.thread_router")
+    @patch("ccgram.handlers.topic_lifecycle.clear_topic_state", new_callable=AsyncMock)
+    @patch("ccgram.handlers.topic_lifecycle.thread_router")
     async def test_skips_unbound_topic(
         self, mock_tr: MagicMock, mock_clear: AsyncMock, _allowed: MagicMock
     ) -> None:
-        from ccgram.bot import topic_closed_handler
+        from ccgram.handlers.topic_lifecycle import topic_closed_handler
 
         mock_tr.get_window_for_thread.return_value = None
 
@@ -57,27 +57,27 @@ class TestTopicClosedHandler:
         mock_tr.unbind_thread.assert_not_called()
         mock_clear.assert_not_called()
 
-    @patch("ccgram.bot.clear_topic_state", new_callable=AsyncMock)
-    @patch("ccgram.bot.thread_router")
+    @patch("ccgram.handlers.topic_lifecycle.clear_topic_state", new_callable=AsyncMock)
+    @patch("ccgram.handlers.topic_lifecycle.thread_router")
     async def test_skips_disallowed_user(
         self, mock_tr: MagicMock, mock_clear: AsyncMock
     ) -> None:
-        from ccgram.bot import topic_closed_handler
+        from ccgram.handlers.topic_lifecycle import topic_closed_handler
 
         update = _make_update()
-        with patch("ccgram.bot.is_user_allowed", return_value=False):
+        with patch("ccgram.config.Config.is_user_allowed", return_value=False):
             await topic_closed_handler(update, MagicMock())
 
         mock_tr.get_window_for_thread.assert_not_called()
         mock_clear.assert_not_called()
 
     @_PATCH_ALLOWED
-    @patch("ccgram.bot.clear_topic_state", new_callable=AsyncMock)
-    @patch("ccgram.bot.thread_router")
+    @patch("ccgram.handlers.topic_lifecycle.clear_topic_state", new_callable=AsyncMock)
+    @patch("ccgram.handlers.topic_lifecycle.thread_router")
     async def test_skips_general_topic(
         self, mock_tr: MagicMock, mock_clear: AsyncMock, _allowed: MagicMock
     ) -> None:
-        from ccgram.bot import topic_closed_handler
+        from ccgram.handlers.topic_lifecycle import topic_closed_handler
 
         update = MagicMock()
         update.effective_user.id = 1
@@ -89,12 +89,12 @@ class TestTopicClosedHandler:
         mock_clear.assert_not_called()
 
     @_PATCH_ALLOWED
-    @patch("ccgram.bot.clear_topic_state", new_callable=AsyncMock)
-    @patch("ccgram.bot.thread_router")
+    @patch("ccgram.handlers.topic_lifecycle.clear_topic_state", new_callable=AsyncMock)
+    @patch("ccgram.handlers.topic_lifecycle.thread_router")
     async def test_skips_no_thread_id(
         self, mock_tr: MagicMock, mock_clear: AsyncMock, _allowed: MagicMock
     ) -> None:
-        from ccgram.bot import topic_closed_handler
+        from ccgram.handlers.topic_lifecycle import topic_closed_handler
 
         update = MagicMock()
         update.effective_user.id = 1
